chore: organise modules, modularise interaction and message events
