+++ src/interfaces/GuildMessage.ts
@@ -1,0 +1,6 @@
+import { Guild, GuildMember, Message } from "discord.js";
+
+export interface GuildMessage extends Message {
+  guild: Guild;
+  member: GuildMember;
+}
